refactor: group standard module files by feature dirs
