+++ projects/ngx-meta/src/standard/src/managers/index.ts
@@ -1,0 +1,13 @@
+export { STANDARD_APPLICATION_NAME_METADATA_PROVIDER } from './standard-application-name-metadata-provider'
+export { STANDARD_AUTHOR_METADATA_PROVIDER } from './standard-author-metadata-provider'
+export { STANDARD_CANONICAL_URL_METADATA_PROVIDER } from './standard-canonical-url-metadata-provider'
+export { STANDARD_DESCRIPTION_METADATA_PROVIDER } from './standard-description-metadata-provider'
+export { STANDARD_GENERATOR_METADATA_PROVIDER } from './standard-generator-metadata-provider'
+export { STANDARD_KEYWORDS_METADATA_PROVIDER } from './standard-keywords-metadata-provider'
+export { STANDARD_LOCALE_METADATA_PROVIDER } from './standard-locale-metadata-provider'
+export { STANDARD_THEME_COLOR_METADATA_PROVIDER } from './standard-theme-color-metadata-provider'
+export {
+  StandardThemeColorMetadata,
+  StandardThemeColorMetadataObject,
+} from './standard-theme-color-metadata'
+export { STANDARD_TITLE_METADATA_PROVIDER } from './standard-title-metadata-provider'
