+++ projects/ngx-meta/src/standard/src/managers/standard-canonical-url-metadata-provider.ts
@@ -1,4 +1,4 @@
-import { makeStandardMetadataProvider } from './make-standard-metadata-provider'
+import { makeStandardMetadataProvider } from '../utils/make-standard-metadata-provider'
 import {
   _GLOBAL_CANONICAL_URL,
   _HEAD_ELEMENT_UPSERT_OR_REMOVE,
