+++ projects/ngx-meta/src/standard/src/managers/standard-generator-metadata-provider.spec.ts
@@ -3,8 +3,8 @@ import { MockProvider } from 'ng-mocks'
 import { enableAutoSpy } from '@/ngx-meta/test/enable-auto-spy'
 import { MetadataSetter, NgxMetaMetaService } from '@davidlj95/ngx-meta/core'
 import { VERSION } from '@angular/core'
-import { Standard } from './standard'
-import { __STANDARD_GENERATOR_METADATA_SETTER_FACTORY } from './standard-generator-metadata-provider'
+import { Standard } from '../types/standard'
+import { STANDARD_GENERATOR_METADATA_SETTER_FACTORY } from './standard-generator-metadata-provider'
 
 describe('Standard generator metadata', () => {
   enableAutoSpy()
@@ -45,7 +45,7 @@ function makeSut(): MetadataSetter<Standard['generator']> {
   TestBed.configureTestingModule({
     providers: [MockProvider(NgxMetaMetaService)],
   })
-  return __STANDARD_GENERATOR_METADATA_SETTER_FACTORY(
+  return STANDARD_GENERATOR_METADATA_SETTER_FACTORY(
     TestBed.inject(NgxMetaMetaService),
   )
 }
