+++ projects/ngx-meta/src/standard/src/managers/standard-keywords-metadata-provider.ts
@@ -1,22 +1,16 @@
-import { makeStandardMetadataProvider } from './make-standard-metadata-provider'
-import { Standard } from './standard'
+import { makeStandardMetadataProvider } from '../utils/make-standard-metadata-provider'
+import { Standard } from '../types/standard'
 import {
   MetadataSetterFactory,
   NgxMetaMetaService,
 } from '@davidlj95/ngx-meta/core'
-import { makeStandardMetaDefinition } from './make-standard-meta-definition'
+import { makeStandardMetaDefinition } from '../utils/make-standard-meta-definition'
 
-/**
- * @internal
- */
-export const __STANDARD_KEYWORDS_METADATA_SETTER_FACTORY: MetadataSetterFactory<
+export const STANDARD_KEYWORDS_METADATA_SETTER_FACTORY: MetadataSetterFactory<
   Standard[typeof KEY]
 > = (metaService: NgxMetaMetaService) => (value) =>
   metaService.set(makeStandardMetaDefinition(KEY), value?.join(','))
 
-/**
- * @internal
- */
 const KEY = 'keywords' satisfies keyof Standard
 
 /**
@@ -25,5 +19,5 @@ const KEY = 'keywords' satisfies keyof Standard
  */
 export const STANDARD_KEYWORDS_METADATA_PROVIDER = makeStandardMetadataProvider(
   KEY,
-  { s: __STANDARD_KEYWORDS_METADATA_SETTER_FACTORY },
+  { s: STANDARD_KEYWORDS_METADATA_SETTER_FACTORY },
 )
