+++ projects/ngx-meta/src/standard/src/managers/standard-theme-color-metadata-provider.spec.ts
@@ -1,8 +1,8 @@
-import { MetadataSetter, NgxMetaMetaService } from '../../core'
-import { Standard } from './standard'
+import { MetadataSetter, NgxMetaMetaService } from '@davidlj95/ngx-meta/core'
+import { Standard } from '../types/standard'
 import { TestBed } from '@angular/core/testing'
 import { MockProvider } from 'ng-mocks'
-import { __STANDARD_THEME_COLOR_METADATA_SETTER_FACTORY } from './standard-theme-color-metadata-provider'
+import { STANDARD_THEME_COLOR_METADATA_SETTER_FACTORY } from './standard-theme-color-metadata-provider'
 import { enableAutoSpy } from '@/ngx-meta/test/enable-auto-spy'
 import { MetaDefinition } from '@angular/platform-browser'
 import { StandardThemeColorMetadataObject } from './standard-theme-color-metadata'
@@ -91,7 +91,7 @@ function makeSut(): MetadataSetter<Standard['themeColor']> {
   TestBed.configureTestingModule({
     providers: [MockProvider(NgxMetaMetaService)],
   })
-  return __STANDARD_THEME_COLOR_METADATA_SETTER_FACTORY(
+  return STANDARD_THEME_COLOR_METADATA_SETTER_FACTORY(
     TestBed.inject(NgxMetaMetaService),
   )
 }
