+++ projects/ngx-meta/src/standard/src/managers/standard-theme-color-metadata.ts
