+++ projects/ngx-meta/src/standard/src/managers/standard-title-metadata-provider.spec.ts
@@ -3,8 +3,8 @@ import { TestBed } from '@angular/core/testing'
 import { MockProvider } from 'ng-mocks'
 import { Title } from '@angular/platform-browser'
 import { MetadataSetter } from '@davidlj95/ngx-meta/core'
-import { Standard } from './standard'
-import { __STANDARD_TITLE_METADATA_SETTER_FACTORY } from './standard-title-metadata-provider'
+import { Standard } from '../types/standard'
+import { STANDARD_TITLE_METADATA_SETTER_FACTORY } from './standard-title-metadata-provider'
 
 describe('Standard title metadata', () => {
   enableAutoSpy()
@@ -43,5 +43,5 @@ function makeSut(): MetadataSetter<Standard['title']> {
   TestBed.configureTestingModule({
     providers: [MockProvider(Title)],
   })
-  return __STANDARD_TITLE_METADATA_SETTER_FACTORY(TestBed.inject(Title))
+  return STANDARD_TITLE_METADATA_SETTER_FACTORY(TestBed.inject(Title))
 }
