+++ projects/ngx-meta/src/standard/src/providers/index.ts
@@ -1,0 +1,2 @@
+export { NgxMetaStandardModule } from './ngx-meta-standard.module'
+export { provideNgxMetaStandard } from './provide-ngx-meta-standard'
