+++ projects/ngx-meta/src/standard/src/providers/ngx-meta-standard.module.ts
