+++ projects/ngx-meta/src/standard/src/types/index.ts
@@ -1,0 +1,2 @@
+export { StandardMetadata } from './standard-metadata'
+export { Standard } from './standard'
